+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - added button to Switch Materials component to force max compression
 - added button to Switch Materials component to force 512x512 resolution
+- changed existing material detection to search for any material with "Quest" in name
 
 # 2.0.0